+++ application/language/english/common_lang.php
@@ -1,0 +1,12 @@
+<?php
+$lang['app.title'] = 'Archmage: Reincarnation from Hell';
+$lang['nav.home'] = 'Home';
+$lang['nav.admin'] = 'Admin';
+$lang['nav.language'] = 'Language';
+$lang['action.create'] = 'Create';
+$lang['action.edit'] = 'Edit';
+$lang['action.delete'] = 'Delete';
+$lang['action.save'] = 'Save';
+$lang['action.cancel'] = 'Cancel';
+$lang['msg.error'] = 'Error';
+$lang['msg.success'] = 'Success';