+++ application/language/english/content_lang.php
@@ -1,0 +1,10 @@
+<?php
+$lang['content.title'] = 'Content Backoffice';
+$lang['content.subtitle'] = 'Units, Spells, Items and Heroes';
+$lang['content.import'] = 'Import CSV/ODS';
+$lang['content.table'] = 'Table';
+$lang['content.file'] = 'File';
+$lang['content.result'] = 'Import result';
+$lang['content.rows_imported'] = 'Rows imported';
+$lang['content.error'] = 'Error';
+$lang['content.back'] = 'Back';