+++ application/language/spanish/battle_lang.php
@@ -1,0 +1,6 @@
+<?php
+$lang['battle.can_attack.denied.below_band'] = 'El objetivo está por debajo de tu banda de ataque';
+$lang['battle.can_attack.denied.above_band'] = 'El objetivo está por encima de tu banda de ataque';
+$lang['battle.counter.available'] = 'El contraataque está disponible';
+$lang['battle.protection.meditation'] = 'Objetivo en Meditación';
+$lang['battle.protection.damage'] = 'Objetivo con Protección de Daño';