+++ application/language/spanish/war_lang.php
@@ -1,0 +1,3 @@
+<?php
+$lang['war.declare.title'] = 'Declarar Guerra';
+$lang['war.scoreboard.title'] = 'Marcador de Guerra';